Move index procedures to an index folder
